requirements.txt
streamlit
python-dotenv
torch
pinecone
langchain
langchain-mistralai
langchain-pinecone
langchain-huggingface
langchain-core
Pillow
sentence-transformers